+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 # 2.0.5(2024-09-xx)
 - input url supports now scheme `file://...` (which is not necessary because file paths are supported). Gzip files are also supported.     
+- sort takes now a sequence for channel values which has higher priority than sort order
 
 # 2.0.4(2024-09-19)
 * if Content type of file download is not set in header, the gzip encoding is checked through magic header.